pyproject.toml
[tool.poetry]
name = "pokemon-proxy"
version = "0.1.0"
description = "A Django REST API that proxies Pokémon abilities"
authors = ["Yasser Azan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
